+++ packages/template-retail-react-app/app/constants.js
@@ -157,5 +157,6 @@ export const SUPPORTED_STORE_LOCATOR_COUNTRIES = [
     }
 ]
 
-export const DEFAULT_STORE_LOCATORY_COUNTRY_CODE = "DE"
-export const DEFAULT_STORE_LOCATORY_POSTAL_CODE = '10178'
+export const DEFAULT_STORE_LOCATOR_COUNTRY_CODE = 'DE'
+export const DEFAULT_STORE_LOCATOR_POSTAL_CODE = '10178'
+export const STORE_LOCATOR_DISTANCE = 100
