+++ packages/template-retail-react-app/app/static/translations/compiled/en-GB.json
@@ -2915,6 +2915,12 @@
       "value": "away"
     }
   ],
+  "store_locator.description.no_locations": [
+    {
+      "type": 0,
+      "value": "Sorry, there are no locations in this area"
+    }
+  ],
   "store_locator.description.or": [
     {
       "type": 0,
@@ -2930,7 +2936,15 @@
   "store_locator.description.viewing_within": [
     {
       "type": 0,
-      "value": "Viewing stores within 100 miles"
+      "value": "Viewing stores within "
+    },
+    {
+      "type": 1,
+      "value": "distance"
+    },
+    {
+      "type": 0,
+      "value": " miles"
     }
   ],
   "store_locator.dropdown.germany": [
@@ -2943,6 +2957,12 @@
     {
       "type": 0,
       "value": "United States"
+    }
+  ],
+  "store_locator.field.placeholder.enter_postal_code": [
+    {
+      "type": 0,
+      "value": "Enter postal code"
     }
   ],
   "store_locator.title": [
